requirements.txt
alabaster==0.7.12
arrow==1.1.1
awscli==1.19.105
Babel==2.9.1
backcall==0.2.0
binaryornot==0.4.4
blis==0.7.4
bokeh==2.3.3
botocore==1.20.105
catalogue==2.0.4
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
colorama==0.4.3
cookiecutter==1.7.3
coverage==5.5
cycler==0.10.0
cymem==2.0.5
decorator==5.0.9
docutils==0.15.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
flake8==3.9.2
funcy==1.16
future==0.18.2
gensim==4.0.1
idna==2.10
imagesize==1.2.0
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
jinja2-time==0.2.0
jmespath==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mccabe==0.6.1
murmurhash==1.0.5
nltk==3.6.2
numexpr==2.7.3
numpy==1.21.0
packaging==20.9
pandas==1.2.5
parso==0.8.2
pathy==0.5.2
pickleshare==0.7.5
Pillow==8.2.0
poyo==0.5.0
preshed==3.0.5
prompt-toolkit==3.0.19
pyasn1==0.4.8
pycodestyle==2.7.0
pydantic==1.7.4
pyflakes==2.3.1
Pygments==2.9.0
pyLDAvis==3.3.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.18.0
python-Levenshtein==0.12.2
python-slugify==5.0.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
rsa==4.7.2
s3transfer==0.4.2
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
smart-open==3.0.0
snowballstemmer==2.1.0
spacy==3.0.6
spacy-legacy==3.0.6
Sphinx==4.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
srsly==2.4.1
text-unidecode==1.3
thinc==8.0.6
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.6
wasabi==0.8.2
wcwidth==0.2.5
